AFP-313: Setup Now deploy of changeset bot
