+++ package.json
@@ -111,11 +111,13 @@
     "lodash.snakecase": "^4.1.1",
     "next": "^8.0.4",
     "next-images": "^1.1.1",
+    "now": "^15.3.0",
     "outdent": "^0.7.0",
     "preconstruct": "^0.0.62",
     "pretty-proptypes": "^0.6.0",
     "prismjs": "^1.15.0",
-    "probot": "^7.2.0",
+    "probot": "^9.2.0",
+    "probot-serverless-now": "^2.0.0",
     "prop-types": "^15.6.2",
     "raw-loader": "^1.0.0",
     "react": "^16.8.4",
