+++ packages/changeset-bot/now.json
@@ -1,0 +1,10 @@
+{
+    "version": 2,
+    "builds": [{ "src": "index.js", "use": "@now/node" }],
+    "env": {
+        "APP_ID": "@app-id",
+        "WEBHOOK_SECRET": "@webhook-secret",
+        "PRIVATE_KEY": "@private-key-base64-encoded"
+    },
+    "routes": [{ "src": "/", "dest": "/index.js" }]
+}
