+++ scripts/deploy_changeset_bot.sh
@@ -2,7 +2,7 @@
 
 cp ../packages/changeset-bot/package.json ../deploy/changeset-bot
 cp ../packages/changeset-bot/yarn.lock ../deploy/changeset-bot
-cp ../packages/changeset-bot/index.js ../deploy/changeset-bot
+cp ../packages/changeset-bot/app.js ../deploy/changeset-bot
 cp ../packages/changeset-bot/.env ../deploy/changeset-bot
 
 cd ../deploy/changeset-bot
@@ -10,7 +10,7 @@ cd ../deploy/changeset-bot
 # Deploy
 docker build -t docker.atl-paas.net/atlassian/changeset-bot:0.0.1 .
 docker push docker.atl-paas.net/atlassian/changeset-bot:0.0.1
-micros service:deploy changeset-bot -f changeset-bot.sd.yml -e stg-east
+micros service:deploy changeset-bot -f changeset-bot.sd.yml
 
 echo "You did it hooray"
